+++ packable/packable/src/packable/prefix/string.rs
@@ -81,12 +81,13 @@ where
 
 impl<B> Packable for StringPrefix<B>
 where
-    B: Bounded + Packable,
+    B: Bounded + Packable<UnpackVisitor = ()>,
     <B::Bounds as TryInto<B>>::Error: fmt::Debug,
     <B as TryFrom<usize>>::Error: fmt::Debug,
     Range<B::Bounds>: Iterator<Item = B::Bounds>,
 {
     type UnpackError = UnpackPrefixError<FromUtf8Error, B::UnpackError>;
+    type UnpackVisitor = ();
 
     #[inline]
     fn pack<P: Packer>(&self, packer: &mut P) -> Result<(), P::Error> {
@@ -102,9 +103,10 @@ where
     #[inline]
     fn unpack<U: Unpacker, const VERIFY: bool>(
         unpacker: &mut U,
+        visitor: &Self::UnpackVisitor,
     ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
         // The length of any dynamically-sized sequence must be prefixed.
-        let len = B::unpack::<_, VERIFY>(unpacker)
+        let len = B::unpack::<_, VERIFY>(unpacker, visitor)
             .map_packable_err(UnpackPrefixError::Prefix)?
             .into();